Add Github Story Points feature.
The GitHub Story Points file has been adapted from
https://github.com/mcomella/github-story-points/blob/master/src/main.js

As the original author, I relicense it under the MPLv2.0 for this project.

The script was modified to compile with TypeScript. I did not do things the
best way: I just got them to work.

Additionally, tsconfig and main.ts were modified to support this feature.